Install and customize Tom Select
https://tom-select.js.org

This may do the trick for now. I was about to spend a bunch of money on
Alpine components (and I may still), but I'm glad not to.

Note to self: the lack of those map files was making `collectstatic`
throw up. There's probably a cleaner way of handling this.