Use Vite to bundle some JS stuff
This took all day. I went down so many rabbit holes. The "hot module
reloading" for CSS is cool, but it comes at the expense of loading CSS
with JS which is not worth it.

So the django-vite plugin is probably not needed. I just need Vite and
its config file.