Handle focus changes from everything to title.
This was super tricky. One of the issues was that I had duplicate IDs
on the status pages. Woof.

https://docs.djangoproject.com/en/5.1/ref/forms/api/#prefixes-for-forms

via:
html - Avoid duplication of form input element ID in Django
https://stackoverflow.com/a/1253364

This discussion helped me figure out timing issues.
https://github.com/alpinejs/alpine/discussions/491

I find it disappointing that hyperscript does so much better of a job
of this on Cassette Nest. Something like…

```
_="on click
    set thisField to the closest .field
    set nextField to the next .field
        within the closest .progressive-field-wrapper
        with wrapping

    toggle .hidden on nextField
    toggle .hidden on thisField

    set nextInput to the first <[x-ref]/> in nextField
    call nextInput.focus()">
```

Verbose, but it's pretty easy to reason about once it's written.

Compare that to this brittle nonsense:

```
toggleFields() {
  this.everything = !this.everything;
  this.$nextTick(() => {
    const inputToFocus = (this.everything) ? this.fieldEverything : this.fieldTitle;
    if (inputToFocus) {
      setTimeout(() => {
        inputToFocus.focus();
      }, 25);
    }
  });
}
```